\name{GeoGenetix-package}
\alias{GeoGenetix-package}
\alias{GeoGenetix}
\docType{package}
\title{Inference and model selection for analysis of geographical
  genetic variation}
\description{MCMC inference and model selection by cross-validation for the analysis of geographical  genetic variation}
\details{
\tabular{ll}{
Package: \tab GeoGenetix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-20\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Filippo Botta and Gilles Guilllot
Who wrote it

Maintainer: Filippo Botta  <filippo.botta@gmail.com>
}
\references{Cf. \code{citation(GeoGenetix)}}

\keyword{ package }
%\seealso{}
%\examples{}
